Major refactor to allow compatibility and new features
