+++ OpenTelemetry/OtInitializationOptions.cs
@@ -12,9 +12,14 @@ namespace Yextly.OpenTelemetry;
 public sealed class OtInitializationOptions
 {
     /// <summary>
-    /// Gets or sets the activity source name used to create operations.
+    /// Gets or sets the pattern to used to register future activity sources.
     /// </summary>
-    public string? ActivitySourceName { get; set; }
+    public string? ActivitySourcePattern { get; set; }
+
+    /// <summary>
+    /// Gets or sets the activity source name used to create operations when the non-generic client is used.
+    /// </summary>
+    public string? DefaultActivitySourceName { get; set; }
 
     /// <summary>
     /// Gets or sets a value indicating whether telemetry is enabled.
