+++ OpenTelemetry/OtTelemetryClient.cs
@@ -4,9 +4,6 @@
 //
 // ==--==
 
-using Microsoft.Extensions.DependencyInjection;
-using OpenTelemetry.Metrics;
-using OpenTelemetry.Trace;
 using System.Diagnostics;
 using Yextly.Telemetry.Abstractions;
 
@@ -15,211 +12,53 @@ namespace Yextly.OpenTelemetry;
 /// <summary>
 /// <see cref="ITelemetryClient" /> implementation backed by <see cref="ActivitySource" />.
 /// </summary>
-public sealed class OtTelemetryClient : ITelemetryClient
+public sealed class OtTelemetryClient : OtTelemetryClientCore
 {
-    private readonly OtImmutableInitializationOptions _options;
-
-    /// <summary>
-    /// Initializes a new instance of the <see cref="OtTelemetryClient" /> class.
-    /// </summary>
-    public OtTelemetryClient(OtImmutableInitializationOptions options, IServiceProvider serviceProvider)
+    /// <inheritdoc />
+    public OtTelemetryClient(OtImmutableInitializationOptions options, IServiceProvider serviceProvider) : base(options, serviceProvider)
     {
         ArgumentNullException.ThrowIfNull(options);
         ArgumentNullException.ThrowIfNull(serviceProvider);
+    }
+}
 
-        // Forces the creation of the Otel pipeline
-        _ = serviceProvider.GetService<TracerProvider>();
-        _ = serviceProvider.GetService<MeterProvider>();
-
-        _options = options;
+/// <summary>
+/// <see cref="ITelemetryClient{T}" /> implementation backed by <see cref="ActivitySource" />.
+/// </summary>
+public sealed class OtTelemetryClient<T> : OtTelemetryClientCore, ITelemetryClient<T> where T : ITelemetryActivitySourceProvider, new()
+{
+    /// <inheritdoc />
+    public OtTelemetryClient(OtImmutableInitializationOptions options, IServiceProvider serviceProvider) : base(CreateOptions(options), serviceProvider)
+    {
+        ArgumentNullException.ThrowIfNull(options);
+        ArgumentNullException.ThrowIfNull(serviceProvider);
     }
 
-    /// <summary>
-    /// Gets the activity source used by this telemetry client.
-    /// </summary>
-    public ActivitySource ActivitySource => _options.ActivitySource;
-
-    /// <inheritdoc />
-    public ITelemetryPropertyBag CreatePropertyBag()
+    [System.Diagnostics.CodeAnalysis.SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "We can't crash in the DI constructor")]
+    private static OtImmutableInitializationOptions CreateOptions(OtImmutableInitializationOptions options)
     {
-        return new OtPropertyBag();
-    }
+        ArgumentNullException.ThrowIfNull(options);
 
-    /// <inheritdoc />
-    public Task<bool> FlushAsync(CancellationToken cancellationToken = default)
-    {
-        return Task.FromResult(true);
-    }
-
-    /// <inheritdoc />
-    public void TrackEvent(string eventName, ITelemetryPropertyBag? properties = default)
-    {
-        Activity? activity = Activity.Current;
-
-        if (activity is null)
+        try
         {
-            using var operation = (OtOperation)TrackOperation(eventName, "event");
-            activity = Activity.Current;
-        }
+            var provider = new T();
 
-        if (activity is not null)
-        {
-            TrackEventCore(activity, eventName, properties);
-        }
-    }
+            // Here we do not handle IsEnabled on purpose: the DI registrar will register a no-op when disabled.
+            var sourceName = provider.SourceName ?? options.ActivitySourceName;
+            var source = new ActivitySource(sourceName);
 
-    /// <inheritdoc />
-    public void TrackException(Exception exception, ITelemetryPropertyBag? properties = default)
-    {
-        ArgumentNullException.ThrowIfNull(exception);
-
-        Activity? activity = Activity.Current;
-
-        if (activity is null)
-        {
-            using var operation = (OtOperation)TrackOperation(exception.GetType().Name, "exception");
-            activity = Activity.Current;
-        }
-
-        if (activity is not null)
-        {
-            TrackExceptionCore(activity, exception, properties);
-        }
-    }
-
-    /// <inheritdoc />
-    public ITelemetryOperation TrackOperation(string operationName, string type)
-    {
-        return TrackOperation(operationName, type, string.Empty, null);
-    }
-
-    /// <inheritdoc />
-    [System.Diagnostics.CodeAnalysis.SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "Many things could go wrong, and we ignore which exception types could be thrown.")]
-    public ITelemetryOperation TrackOperation(string operationName, string type, string operationId, string? parentOperationId = default)
-    {
-        ActivityContext? parentContext = null;
-
-        // Convert legacy Application Insights identifiers -> OpenTelemetry context
-        if (!string.IsNullOrWhiteSpace(operationId))
-        {
-            try
+            var newOptions = options with
             {
-                var traceId = ActivityTraceId.CreateFromString(operationId.AsSpan());
+                ActivitySourceName = sourceName,
+                ActivitySource = source,
+            };
 
-                var spanId = !string.IsNullOrWhiteSpace(parentOperationId)
-                    ? ActivitySpanId.CreateFromString(parentOperationId.AsSpan())
-                    : ActivitySpanId.CreateRandom();
-
-                parentContext = new ActivityContext(traceId, spanId, ActivityTraceFlags.Recorded);
-            }
-            catch
-            {
-                // Ignore invalid IDs (avoid crashing telemetry)
-                parentContext = null;
-            }
+            return newOptions;
         }
-
-        var kind = MapActivityKind(type);
-
-        var tags = new List<KeyValuePair<string, object?>>
+        catch
         {
-            new("dependency.type", type),
-        };
-
-        Activity? activity = StartActivity(operationName, kind, parentContext, tags);
-
-        return new OtOperation(this, activity);
-    }
-
-    [System.Diagnostics.CodeAnalysis.SuppressMessage("Performance", "CA1822:Mark members as static", Justification = "Required internally non static by design.")]
-    internal void TrackEventCore(Activity activity, string eventName, ITelemetryPropertyBag? properties)
-    {
-        ArgumentNullException.ThrowIfNull(activity);
-
-        activity.AddEvent(new ActivityEvent(eventName, tags: CreateTags(properties)));
-    }
-
-    [System.Diagnostics.CodeAnalysis.SuppressMessage("Performance", "CA1822:Mark members as static", Justification = "Required internally non static by design.")]
-    internal void TrackExceptionCore(Activity activity, Exception exception, ITelemetryPropertyBag? properties)
-    {
-        ArgumentNullException.ThrowIfNull(activity);
-        ArgumentNullException.ThrowIfNull(exception);
-
-        activity.SetStatus(ActivityStatusCode.Error, exception.Message);
-
-        var tags = new ActivityTagsCollection
-        {
-            { "exception.type", exception.GetType().FullName },
-            { "exception.message", exception.Message },
-            { "exception.stacktrace", exception.ToString() },
-        };
-
-        if (properties is OtPropertyBag propertyBag)
-        {
-            foreach ((string key, string value) in propertyBag.Data)
-            {
-                tags[key] = value;
-            }
+            // In case of bad things, we don't want to smash the whole application with an exception hard to track.
+            return options;
         }
-
-        activity.AddEvent(new ActivityEvent("exception", tags: tags));
-    }
-
-    private static ActivityTagsCollection? CreateTags(ITelemetryPropertyBag? properties)
-    {
-        if (properties is not OtPropertyBag propertyBag || propertyBag.Data.Count == 0)
-        {
-            return null;
-        }
-
-        var tags = new ActivityTagsCollection();
-
-        foreach ((string key, string value) in propertyBag.Data)
-        {
-            tags[key] = value;
-        }
-
-        return tags;
-    }
-
-    private static ActivityKind MapActivityKind(string type)
-    {
-        return type switch
-        {
-            "http" => ActivityKind.Client,
-            "db" => ActivityKind.Client,
-            "messaging" => ActivityKind.Producer,
-            "server" => ActivityKind.Server,
-            _ => ActivityKind.Internal
-        };
-    }
-
-    private Activity? StartActivity(string operationName, ActivityKind kind, ActivityContext? parentContext, IEnumerable<KeyValuePair<string, object?>>? tags)
-    {
-        var activity = parentContext is null
-            ? ActivitySource.StartActivity(operationName, kind, parentId: null, tags: tags)
-            : ActivitySource.StartActivity(operationName, kind, parentContext.Value, tags);
-
-        if (activity is null)
-        {
-            activity = new Activity(operationName);
-
-            if (parentContext is not null)
-            {
-                activity.SetParentId(parentContext.Value.TraceId, parentContext.Value.SpanId, parentContext.Value.TraceFlags);
-            }
-
-            //if (tags is not null)
-            //{
-            //    foreach ((string key, object? value) in tags)
-            //    {
-            //        activity.SetTag(key, value);
-            //    }
-            //}
-
-            activity.Start();
-        }
-
-        return activity;
     }
 }
