+++ TelemetryAbstractions/ITelemetryClient.cs
@@ -6,8 +6,9 @@
 
 namespace Yextly.Telemetry.Abstractions
 {
+
     /// <summary>
-    /// Represents the abstraction client.
+    /// Represents the abstract telemetry client.
     /// </summary>
     public interface ITelemetryClient : ITelemetryCommonOperation
     {
@@ -15,28 +16,36 @@ namespace Yextly.Telemetry.Abstractions
         /// Asynchronously flushes the telemetry buffers.
         /// </summary>
         /// <remarks>You need to manually flush the buffers when the application exits</remarks>
-        /// <param name="cancellationToken">The token to monitor for cancellation requests The default value is <see cref="CancellationToken.None"/>.</param>
+        /// <param name="cancellationToken">The token to monitor for cancellation requests The default value is <see cref="CancellationToken.None" />.</param>
         /// <returns></returns>
         Task<bool> FlushAsync(CancellationToken cancellationToken = default);
 
         /// <summary>
-        /// Tracks an operation and returns an instance implementing <see cref="ITelemetryOperation"/> which must be kept alive for the whole duration of the operation that is being tracked.
+        /// Tracks an operation and returns an instance implementing <see cref="ITelemetryOperation" /> which must be kept alive for the whole duration of the operation that is being tracked.
         /// </summary>
         /// <param name="operationName">The name of the operation.</param>
         /// <param name="type">The type of the operation to track.</param>
-        /// <remarks>The meaning of <paramref name="type"/> depends on the precise implementation used.</remarks>
+        /// <remarks>The meaning of <paramref name="type" /> depends on the precise implementation used.</remarks>
         /// <returns></returns>
         ITelemetryOperation TrackOperation(string operationName, string type);
 
         /// <summary>
-        /// Tracks an operation and returns an instance implementing <see cref="ITelemetryOperation"/> which must be kept alive for the whole duration of the operation that is being tracked.
+        /// Tracks an operation and returns an instance implementing <see cref="ITelemetryOperation" /> which must be kept alive for the whole duration of the operation that is being tracked.
         /// </summary>
         /// <param name="operationName">The name of the operation.</param>
         /// <param name="type">The type of the operation to track.</param>
         /// <param name="operationId">The unique identifier of the operation.</param>
         /// <param name="parentOperationId">The parent operation identifier.</param>
-        /// <remarks>The meaning of <paramref name="type"/> depends on the precise implementation used.</remarks>
+        /// <remarks>The meaning of <paramref name="type" /> depends on the precise implementation used.</remarks>
         /// <returns></returns>
         ITelemetryOperation TrackOperation(string operationName, string type, string operationId, string? parentOperationId = null);
+    }
+
+    /// <summary>
+    /// Represents the abstract telemetry client.
+    /// </summary>
+    /// <typeparam name="T">The type used for metadata inference.</typeparam>
+    public interface ITelemetryClient<T> : ITelemetryClient where T : ITelemetryActivitySourceProvider, new()
+    {
     }
 }
